refactor code base to prepare for additional autoscaling services (e.g., Kinesis)
