+++ test/test_emr_fetch_cluster_info.py
