+++ themis/model/aws_model.py
@@ -1,0 +1,4 @@
+
+class AwsObject(object):
+    def __init__(self, id):
+        self.id = id
