+++ themis/model/emr_model.py
@@ -1,0 +1,6 @@
+from themis.model.aws_model import *
+
+
+class EmrCluster(AwsObject):
+    def __init__(self, id):
+        super(EmrCluster, self).__init__(id)
