+++ themis/scaling/api.py
@@ -12,6 +12,7 @@ from themis.scaling import server
 from themis.constants import *
 from themis.util import common, monitoring, aws_common, aws_pricing
 from themis.util.aws_common import INSTANCE_GROUP_TYPE_TASK
+from themis.scaling import emr_scaling
 
 root_path = os.path.dirname(os.path.realpath(__file__))
 web_dir = root_path + '/../web/'
@@ -19,7 +20,7 @@ web_dir = root_path + '/../web/'
 app = Flask('app', template_folder=web_dir)
 app.root_path = root_path
 
-CLUSTERS = server.CLUSTERS
+CLUSTERS = emr_scaling.CLUSTERS
 
 
 @app.route('/swagger.json')
