+++ themis/util/exceptions.py
@@ -1,0 +1,4 @@
+
+class ConnectivityException(Exception):
+    def __init__(self, message=None):
+        super(Exception, self).__init__(message)
