+++ themis/web/index.html
@@ -25,13 +25,14 @@
                 <img src="img/scale_white.png" style="width: 30px; height: 30px; padding: 2px; margin-right: 10px;"/></span>
           </h1>
           <ul class="aui-nav">
-            <li><a id="data-link" href="/#/clusters/list">Clusters</a></li>
+            <li><a id="data-link" href="/#/emr/list">EMR</a></li>
+            <li><a id="data-link" href="/#/kinesis/list">Kinesis</a></li>
             <li><a id="data-link" href="/#/config">Settings</a></li>
           </ul>
         </div>
         <div class="aui-header-secondary">
           <ul class="aui-nav">
-            <li><span style="display: block; line-height: 1; padding: 13px 10px;">EMR Cluster Manager</span></li>
+            <li><span style="display: block; line-height: 1; padding: 13px 10px;">Themis Autoscaler</span></li>
           </ul>
         </div>
       </div>
@@ -79,9 +80,11 @@ AJS.tablessortable.setTableSortable($(".aui-table-sortable"));
 </script>
 
 <script src="js/app.js"></script>
-<script src="views/clusters.js"></script>
-<script src="views/clusters.list.js"></script>
-<script src="views/clusters.details.js"></script>
+<script src="views/emr.js"></script>
+<script src="views/emr.list.js"></script>
+<script src="views/emr.details.js"></script>
+<script src="views/kinesis.js"></script>
+<script src="views/kinesis.list.js"></script>
 <script src="views/config.js"></script>
 
 </body>
