+++ themis/web/views/kinesis.html
@@ -1,0 +1,2 @@
+<div ui-view="list"></div>
+<div ui-view="details"></div>
