+++ themis/web/views/kinesis.js
@@ -1,0 +1,10 @@
+(function () {
+  'use strict';
+
+  var app = angular.module('app');
+
+  app.controller('kinesisCtrl', function($scope) {
+    
+  });
+
+})();
