+++ themis/web/views/kinesis.list.html
@@ -1,0 +1,19 @@
+<section id="content">
+
+<header class="aui-page-header">
+  <div class="aui-page-header-inner">
+  <div class="aui-page-header-main">
+    <h1 style="font-size: 24px" id="heading-title">Kinesis Streams</h1>
+  </div>
+  </div>
+</header>
+
+<div class="aui-page-panel">
+  <div class="aui-page-panel-inner">
+    <section class="aui-page-panel-content">
+      Coming soon...
+    </section>
+  </div>
+</div>
+
+</section>
