+++ themis/web/views/kinesis.list.js
@@ -1,0 +1,10 @@
+(function () {
+  'use strict';
+
+  var app = angular.module('app');
+
+  app.controller('kinesisListCtrl', function($scope, restClient, appConfig) {
+    
+  });
+
+})();
